4 New Items, changes to SIF
A backpack item is added, this is a virtual container, a new callback
was added to SIF/Item OnItemCreate which is used to create a matching
container with the backpack to store items in. An OnItemDestroy callback
is on the todo list.

The briefcase has similar functionality to the backpack but it cannot be
worn thus the extra storage cannot be used at the same time as picking
up other items. Wheras when the backpack is used with F it is no longer
'held' by the player so he can pick up items freely.

A shield has been added, it can be placed with F to act as a shield,
also in the main game script if the player is shot from the side he is
holding the shield, the damage is reduced somewhat.

Lastly a wrench has been added that can be used to fix broken vehicles,
I plan on adding more tools soon though and restricting their abilities
somewhat.