SIF and SSS update
Re-worked interface functions for SIF/Item, removed various functions
that returned data for items that was in fact data related to the item
type. These functions are now renamed and require an item type as the
input (which can be retreived with GetItemType)

Added/improved documentation for:
SIF/Button
SIF/Door
SIF/Item
SIF/Inventory

With better distinguishment of Core, Internal and Interface functions.

SIF/Inventory now allows users to add custom options to the item options
menu and thus the dependancy on SIF/Craft is removed as that script now
uses this new system (Which is mainly why I added it)

Map scripts added to separate the entitied created in different areas,
there are now 6 areas: the 3 cities, 2 country areas (red county and
flint county) and the desert (bone c./tierra r.)

SIF will now have it's own repo and this will purely be for SFFA and
maybe SSS updates.