Compatibility Update
Due to changes in SIF and some other things, here's a quick update that
fixes a lot of those things.

Removed rbits as it's not really used much and conflicts with y_bit.

Changed GeoIP plugin to one with a working link and no plugin required,
updated player country display code.